backend/requirements.txt
fastapi==0.115.11
uvicorn
openai==1.64.0
faiss-cpu==1.10.0
numpy==2.0.2
pydantic==2.10.6
pytest==8.3.5
python-dotenv==1.0.1
pydantic-settings==2.8.1